Distributed vma combined with global monitor
I. Distributed vma:
  1. Config:
  - Both processor nodes and memory nodes:
    - Enable CONFIG_DISTRIBUTED_VMA
    - Choose CONFIG_VM_GRANULARITY_ORDER
  - Processor nodes:
    - Enable CONFIG_DISTRIBUTED_VMA_PROCESSOR
  - Memory nodes:
    - Enable CONFIG_DISTRIBUTED_VMA_MEMORY
    - Choose CONFIG_MEM_NR_NODES

II. Global Monitors:
  1. Config: (linux_modules/monitor/include/monitor_config.h)
  - global monitor dispatcher:
    - Choose MAX_RXBUF_SIZE
  - global processor monitor:
    - Choose PROCESSOR_NODE_COUNT
    - Fill pnode_nids array with node ids
  - global memory monitor:
    - Choose MEMORY_NODE_COUNT
    - Fill mnode_nids with node ids
    - Enable MNODES_STATUS_REQUEST
      (if you want to polling physical memory from each memory nodes)
    - Disable ROUND_ROBIN_CHOOSE
      (Enable it will cause round robin memory request allocation)

  2. Usage:
  - modules installation:
  navigate to linux_modules directory and make help for further info
  - start a new process:
  echo "exec=<target_files arguments>" > /proc/lego

  3. Print:
  use dmesg

+++ include/memory/mm.h
@@ -96,8 +96,8 @@ struct vma_tree {
 					   potential grow of MAP_FIXED request
 					   like brk */
 	/* fields below mainly serves lego_mm_struct.node_map */
-	__u64 max_gap;			/* max gap of corresponding range */
-	__u64 mnode;
+	unsigned long max_gap;		/* max gap of corresponding range */
+	int mnode;
 	struct list_head list;
 };
 #endif
